Refactor skybox and texture handling for IBL pipeline
Replaces raw OpenGL texture management in Skybox with a new Texture class and builder pattern, improving resource safety and clarity. Updates all IBL resource generation and rendering code to use smart pointers and the new Texture API. Removes legacy comments and fixes, streamlining shader and resource initialization. Updates related shader and source files to remove 'CRITICAL FIX' comments and clarify logic. Adds documentation guides for texture refactoring.

+++ src/AudioPlayer.h
@@ -108,7 +108,7 @@ public:
         Mix_Volume(channel, vol);
 
         // Compute the angle (in degrees) from listener to source.
-        float angle = std::atan2f(diff.y, diff.x) * 180.0f / M_PI;
+        float angle = std::atan2f(diff.y, diff.x) * 180.0f / static_cast<float>(M_PI);
         float relativeAngle = angle - listenerAngle;
         while (relativeAngle < 0) {
             relativeAngle += 360.0f;
